# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campus Faso - Gestion du LMD",
        "description": "Documentation générée par Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/niveaux": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Récupérer la liste de tous les niveaux",
                "operationId": "667f5a247cdd84268551776de804cb74",
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/niveaux/{id}": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Récupérer un niveau par son ID",
                "operationId": "49cddd711f9136a5cc6e92d141a65bd2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du niveau",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "libelle": {
                                            "type": "string",
                                            "example": "Licence 1"
                                        },
                                        "abreviation": {
                                            "type": "string",
                                            "example": "L1"
                                        },
                                        "filiere_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "accessible": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "statut": {
                                            "type": "string",
                                            "example": "active"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Niveau non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/niveaux/accessibles": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Récupérer la liste des niveaux accessibles",
                "operationId": "b00d2606eeef1fa51f96901b771444d7",
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/niveaux/{niveau}/diplomes": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Récupérer la liste des diplômes requis pour un niveau",
                "operationId": "836666beaa03ddd07c364840fa2438e3",
                "parameters": [
                    {
                        "name": "niveau",
                        "in": "path",
                        "description": "ID du niveau",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Baccalauréat"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "BAC"
                                            },
                                            "serie": {
                                                "type": "string",
                                                "example": "C"
                                            },
                                            "specialite": {
                                                "type": "string",
                                                "example": "Mathématiques"
                                            },
                                            "option": {
                                                "type": "string",
                                                "example": "Sciences"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Niveau non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/niveaux/filiere/{filiere_id}": {
            "get": {
                "tags": [
                    "Niveaux"
                ],
                "summary": "Récupérer les niveaux d'une filière",
                "operationId": "733341a683a0094e97bd7c764eab9064",
                "parameters": [
                    {
                        "name": "filiere_id",
                        "in": "path",
                        "description": "ID de la filière",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "filiere_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}